Fjerne parallell støtte for long og UUID
Vi gjør en big bang-switch allikevel, og trenger derfor ikke å håndtere
oppgavereferanse som long og UUID samtidig.

Flytte tester av godkjennings-APIet fra RestApiTest til egen testklasse
for VedtaksperiodeApi. Trenger ikke å bruke E2E-testing for alt
funksjonalitet i appen.

+++ src/main/kotlin/no/nav/helse/mediator/kafka/HendelseMediator.kt
@@ -297,8 +297,8 @@ internal class HendelseMediator(
     }
 
     internal fun håndter(godkjenningDTO: GodkjenningDTO, epost: String, oid: UUID) {
-        val contextId = oppgaveDao.finnContextId(godkjenningDTO.oppgavereferanse.toLong())
-        val hendelseId = oppgaveDao.finnHendelseId(godkjenningDTO.oppgavereferanse.toLong())
+        val contextId = oppgaveDao.finnContextId(godkjenningDTO.oppgavereferanse)
+        val hendelseId = oppgaveDao.finnHendelseId(godkjenningDTO.oppgavereferanse)
         val fødselsnummer = personDao.finnFødselsnummer(hendelseId)
         val godkjenningMessage = JsonMessage.newMessage(
             standardfelter("saksbehandler_løsning", fødselsnummer).apply {
